split into multiple files, add proper error handling for user user dirs
